requirements.txt
# ImgStegGAN V1.0.0 Dependencies
# Requires PyTorch 2.x

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0

# Numerical Computing
numpy>=1.21.0

# Error Correction Coding
reedsolo>=1.0.0

# Progress Bar
tqdm>=4.64.0

# Web Framework
flask>=2.0.0
flask-cors>=3.0.0

# Transformers for Qwen Integration
transformers>=4.30.0
